+++ dask/tests/test_base.py
@@ -234,6 +234,7 @@ def test_replace_name_in_keys():
 class Tuple(DaskMethodsMixin):
     __slots__ = ("_dask", "_keys")
     __dask_scheduler__ = staticmethod(dask.threaded.get)
+    __dask_optimize__ = None
 
     def __init__(self, dsk, keys):
         self._dask = dsk
@@ -606,16 +607,6 @@ def inc_to_dec(dsk, keys):
     return dsk
 
 
-def test_optimizations_keyword():
-    x = dask.delayed(inc)(1)
-    assert x.compute() == 2
-
-    with dask.config.set(optimizations=[inc_to_dec]):
-        assert x.compute() == 0
-
-    assert x.compute() == 2
-
-
 def test_optimize():
     x = dask.delayed(inc)(1)
     y = dask.delayed(inc)(x)
@@ -631,18 +622,6 @@ def test_optimize():
 
     # Computationally equivalent
     assert dask.compute(x2, y2, z2) == dask.compute(x, y, z)
-
-    # Applying optimizations before compute and during compute gives
-    # same results. Shows optimizations are occurring.
-    sols = dask.compute(x, y, z, optimizations=[inc_to_dec])
-    x3, y3, z3 = optimize(x, y, z, optimizations=[inc_to_dec])
-    assert dask.compute(x3, y3, z3) == sols
-
-    # Optimize respects global optimizations as well
-    with dask.config.set(optimizations=[inc_to_dec]):
-        x4, y4, z4 = optimize(x, y, z)
-    for a, b in zip([x3, y3, z3], [x4, y4, z4]):
-        assert dict(a.dask) == dict(b.dask)
 
 
 def test_optimize_nested():
@@ -866,7 +845,8 @@ def test_optimize_None():
     y = x[:9][1:8][::2] + 1  # normally these slices would be fused
 
     def my_get(dsk, keys):
-        assert dsk == dict(y.dask)  # but they aren't
+        # but they aren't. +1 for the finalize task
+        assert len(dsk) == len(y.dask) + 1
         return dask.get(dsk, keys)
 
     with dask.config.set(array_optimize=None, scheduler=my_get):
@@ -875,7 +855,7 @@ def test_optimize_None():
 
 def test_scheduler_keyword():
     def schedule(dsk, keys, **kwargs):
-        return [[123]]
+        return [123]
 
     named_schedulers["foo"] = schedule